Legacy trainer
Should use Tensorflow or something else as a learning model. This
handmade Q learning could be plaqued with a lot of mistakes, errors and
something I don't know.

This branch and this commit will serve as my memento that once upon a
time, I tried to build this thing once. :wave: